+++ tmp.R
@@ -1,21 +1,6 @@
 # TODO
 
-# Pull abundanceperquad() out from tt_test(), wrap it into n_by_idx() and feed 
-# it as an argument. n_by_idx() can then change once I develop a better version.
-
-
-# Fix errors caused by removing tidyr
-
-# Rewrite documentation to not say that data is long but to point that it can
-# be made long with gather().
-
-
 # Test that tt_test() fails with informative messages. 
-
-# Update vignette to add wrappers tt_test() and tt_gather().
-# Update functions index on fgeo
-# update README.
-
 # Add messages to these functions to report the extracted values.
 # Rename extract_plotdim?
 # Rename extract_gridsize?