+++ NEWS.md
@@ -1,6 +1,6 @@
-# soilkrig 0.0.0.9001
+# fgeo.habitat 0.0.0.9002
 
-* New function `tt_test()` and friends.
+* New function `tt_test()` and friends, by Sabrinna Russo et al.
 * Rename to __fgeo.habitat__
 
 # soilkrig 0.0.0.9000